moved out expressServer and fixed type color in mockdata
